+++ VitaliyKuroeDov/package.json
@@ -16,9 +16,11 @@
     "connect-mongo": "^3.2.0",
     "consolidate": "^0.15.1",
     "cookie-parser": "^1.4.5",
+    "cors": "^2.8.5",
     "express": "^4.17.1",
     "express-session": "^1.17.1",
     "handlebars": "^4.7.6",
+    "jsonwebtoken": "^8.5.1",
     "mongoose": "^5.9.26",
     "node-fetch": "^2.6.0",
     "node-sound-effects": "^1.1.1",
